pyproject.toml
[tool.poetry]
name = "creonit-test"
version = "0.1.0"
description = ""
authors = ["Artem Merkulov <a.merkulov@napoleonit.com>"]
readme = "README.md"
packages = [{include = "creonit_test"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
envparse = "^0.2.0"
django-treebeard = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
